+++ foo_discord_rich/fb2k/advanced_config.h
@@ -8,7 +8,5 @@ namespace drp::config::advanced
 extern qwr::fb2k::AdvConfigBool_MT enableDebugLog;
 extern qwr::fb2k::AdvConfigBool_MT logWebRequests;
 extern qwr::fb2k::AdvConfigBool_MT logWebResponses;
-extern qwr::fb2k::AdvConfigBool_MT logUploaderCmds;
-extern qwr::fb2k::AdvConfigBool_MT logUploaderOutput;
 
 } // namespace drp::config::advanced